fix: satisfy pkg/service backend-size limits for PR CI
Extract runtime factory instantiation from buildRuntimeBundle and split
oversized service test files so backendsizecheck passes on the PR head.
